+++ common/src/main/java/eus/ibai/family/recipes/food/event/EventStoreHealthContributor.java
@@ -1,6 +1,6 @@
 package eus.ibai.family.recipes.food.event;
 
-import eus.ibai.family.recipes.food.health.ComponentHealthContributor;
+import eus.ibai.family.recipes.food.health.AbstractComponentHealthContributor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
@@ -15,22 +15,13 @@ import javax.sql.DataSource;
 @Slf4j
 @Component
 @ConditionalOnProperty(prefix = "management.health.event-store", name = "enabled", havingValue = "true")
-public class EventStoreHealthContributor implements ComponentHealthContributor {
-
-    public static final String COMPONENT_NAME = "event-store";
+public class EventStoreHealthContributor extends AbstractComponentHealthContributor {
 
     private final DataSourceHealthIndicator decoratedIndicator;
 
-    private final Long interval;
-
     public EventStoreHealthContributor(@Qualifier("event-store-db") DataSource dataSource, @Value("${management.health.event-store.interval:300}") long interval) {
+        super("event-store", interval);
         this.decoratedIndicator = new DataSourceHealthIndicator(dataSource);
-        this.interval = interval;
-    }
-
-    @Override
-    public String getComponentName() {
-        return COMPONENT_NAME;
     }
 
     @Override
@@ -38,10 +29,5 @@ public class EventStoreHealthContributor implements ComponentHealthContributor {
         return Mono.just(decoratedIndicator.health())
                 .map(health -> Health.status(health.getStatus()).build())
                 .doOnNext(health -> log.trace("Received event store health response: {}", health.getStatus()));
-    }
-
-    @Override
-    public long getInterval() {
-        return interval;
     }
 }
