+++ write-model/src/main/java/eus/ibai/family/recipes/food/wm/infrastructure/config/AxonConfig.java
@@ -23,7 +23,6 @@ import org.axonframework.extensions.springcloud.commandhandling.mode.AcceptAllCo
 import org.axonframework.extensions.springcloud.commandhandling.mode.RestCapabilityDiscoveryMode;
 import org.axonframework.serialization.json.JacksonSerializer;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.web.client.RestTemplateBuilder;
