+++ write-model/src/main/java/eus/ibai/family/recipes/food/wm/infrastructure/config/HealthIndicatorConfig.java
@@ -1,0 +1,17 @@
+package eus.ibai.family.recipes.food.wm.infrastructure.config;
+
+import eus.ibai.family.recipes.food.health.ComponentHealthIndicator;
+import eus.ibai.family.recipes.food.health.HealthCache;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class HealthIndicatorConfig {
+
+    @Bean("service-registry")
+    @ConditionalOnProperty(prefix = "management.health.service-registry", name = "enabled", havingValue = "true")
+    public ComponentHealthIndicator serviceRegistryHealthIndicator(HealthCache healthCache) {
+        return new ComponentHealthIndicator("service-registry", healthCache);
+    }
+}
