+++ write-model/src/main/java/eus/ibai/family/recipes/food/wm/infrastructure/constraint/PropertyNameConstraintRepository.java
@@ -1,7 +1,6 @@
 package eus.ibai.family.recipes.food.wm.infrastructure.constraint;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
