+++ write-model/src/main/java/eus/ibai/family/recipes/food/wm/infrastructure/health/ServiceRegistryHealthContributor.java
@@ -1,6 +1,6 @@
 package eus.ibai.family.recipes.food.wm.infrastructure.health;
 
-import eus.ibai.family.recipes.food.health.ComponentHealthContributor;
+import eus.ibai.family.recipes.food.health.AbstractComponentHealthContributor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.actuate.health.Health;
@@ -12,22 +12,13 @@ import reactor.core.publisher.Mono;
 @Slf4j
 @Component
 @ConditionalOnProperty(prefix = "management.health.service-registry", name = "enabled", havingValue = "true")
-public class ServiceRegistryHealthContributor implements ComponentHealthContributor {
-
-    private static final String COMPONENT_NAME = "serviceRegistry";
+public class ServiceRegistryHealthContributor extends AbstractComponentHealthContributor {
 
     private final DiscoveryClient discoveryClient;
 
-    private final long interval;
-
     public ServiceRegistryHealthContributor(DiscoveryClient discoveryClient, @Value("${management.health.service-registry.interval:300}") long interval) {
+        super("service-registry", interval);
         this.discoveryClient = discoveryClient;
-        this.interval = interval;
-    }
-
-    @Override
-    public String getComponentName() {
-        return COMPONENT_NAME;
     }
 
     @Override
@@ -39,10 +30,5 @@ public class ServiceRegistryHealthContributor implements ComponentHealthContribu
             log.error("Service registry health failed.", e);
             return Mono.just(Health.down().build());
         }
-    }
-
-    @Override
-    public long getInterval() {
-        return interval;
     }
 }
