+++ write-model/src/main/resources/logback.xml
@@ -13,7 +13,7 @@
     <logger name="eus.ibai" level="DEBUG"/>
     <logger name="org.flywaydb" level="INFO"/>
     <logger name="org.springframework" level="INFO"/>
-    <logger name="org.springframework.web" level="INFO"/>
+    <logger name="org.springframework.web.server.adapter.HttpWebHandlerAdapter" level="INFO"/>
     <logger name="org.springframework.r2dbc.core.DefaultDatabaseClient" level="INFO"/>
     <logger name="io.r2dbc.postgresql.PARAM" level="INFO"/>
     <logger name="io.r2dbc.postgresql.QUERY" level="INFO"/>
