+++ write-model/src/test/java/eus/ibai/family/recipes/food/wm/application/controller/DistributedCommandBusControllerIT.java
@@ -137,7 +137,7 @@ class DistributedCommandBusControllerIT {
     }
 
     @DynamicPropertySource
-    public static void setDatasourceProperties(final DynamicPropertyRegistry registry) {
+    public static void setDynamicProperties(final DynamicPropertyRegistry registry) {
         registry.add("spring.cloud.zookeeper.connect-string", () -> "localhost:" + zookeeperContainer.getHttpPort());
     }
 }
