+++ write-model/src/test/java/eus/ibai/family/recipes/food/wm/infrastructure/health/ServiceRegistryHealthIT.java
@@ -43,7 +43,7 @@ class ServiceRegistryHealthIT {
     }
 
     @DynamicPropertySource
-    private static void setDatasourceProperties(final DynamicPropertyRegistry registry) {
+    private static void setDynamicProperties(final DynamicPropertyRegistry registry) {
         registry.add("spring.cloud.zookeeper.connect-string", () -> "localhost:" + zookeeperContainer.getHttpPort());
     }
 }
